+++ src/Runner.php
@@ -194,12 +194,14 @@ class Runner
         $container->addServiceProvider(\Robo\Task\Archive\loadTasks::getArchiveServices());
         $container->addServiceProvider(\Robo\Task\Assets\loadTasks::getAssetsServices());
         $container->addServiceProvider(\Robo\Task\Base\loadTasks::getBaseServices());
+        $container->addServiceProvider(\Robo\Task\Npm\loadTasks::getNpmServices());
         $container->addServiceProvider(\Robo\Task\Bower\loadTasks::getBowerServices());
+        $container->addServiceProvider(\Robo\Task\Gulp\loadTasks::getGulpServices());
         $container->addServiceProvider(\Robo\Task\Composer\loadTasks::getComposerServices());
         $container->addServiceProvider(\Robo\Task\Development\loadTasks::getDevelopmentServices());
         $container->addServiceProvider(\Robo\Task\Docker\loadTasks::getDockerServices());
         $container->addServiceProvider(\Robo\Task\File\loadTasks::getFileServices());
-        $container->addServiceProvider(\Robo\Task\FileSystem\loadTasks::getFileSystemServices());
+        $container->addServiceProvider(\Robo\Task\Filesystem\loadTasks::getFilesystemServices());
         $container->addServiceProvider(\Robo\Task\Remote\loadTasks::getRemoteServices());
         $container->addServiceProvider(\Robo\Task\Testing\loadTasks::getTestingServices());
         $container->addServiceProvider(\Robo\Task\Vcs\loadTasks::getVcsServices());