+++ src/Task/Filesystem/DeleteDir.php
@@ -1,5 +1,5 @@
 <?php
-namespace Robo\Task\FileSystem;
+namespace Robo\Task\Filesystem;
 
 use Robo\Common\ResourceExistenceChecker;
 use Robo\Result;