+++ src/Task/Filesystem/MirrorDir.php
@@ -1,5 +1,5 @@
 <?php
-namespace Robo\Task\FileSystem;
+namespace Robo\Task\Filesystem;
 
 use Robo\Result;